# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  domain.Events:
    properties:
      applicationclosingdate:
        type: string
      applicationlink:
        type: string
      approved:
        type: boolean
      archived:
        type: boolean
      createdat:
        type: string
      cusatonly:
        type: boolean
      eventdate:
        type: string
      eventid:
        type: integer
      eventpic:
        type: string
      location:
        type: string
      longdiscription:
        type: string
      maxapplications:
        type: integer
      online:
        type: boolean
      organizername:
        type: string
      paid:
        type: boolean
      sex:
        type: string
      shortdiscription:
        type: string
      subevents:
        type: string
      title:
        maxLength: 50
        minLength: 2
        type: string
      websitelink:
        type: string
    required:
    - eventdate
    - eventpic
    - organizername
    - title
    type: object
...
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: sethukumarj.76@gmail.com
    name: API Support
    url: sethukumarj.com
  description: This is an Events Radar project. You can visit the GitHub repository
    at https://github.com/SethukumarJ/Events_Radar_Developement
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go + Gin Radar API
  version: "1.0"
paths:
...
  /admin/create-event:
    post:
      operationId: Create event from admin
      parameters:
      - description: Create event
        in: body
        name: CreateEvent
        required: true
        schema:
          $ref: '#/definitions/domain.Events'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create event
      tags:
      - Admin
...
  /organization/create-event:
    post:
      operationId: Create event from organization
      parameters:
      - description: organizationName
        in: query
        name: organizationName
        required: true
        type: string
      - description: Create event
        in: body
        name: CreateEvent
        required: true
        schema:
          $ref: '#/definitions/domain.Events'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create event by organization
      tags:
      - Organization
  /organization/delete-event:
    delete:
      operationId: Delete event
      parameters:
      - description: 'Title: '
        in: query
        name: title
        required: true
        type: string
      - description: 'organizationName: '
        in: query
        name: organizationName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: delete event
      tags:
      - Organization
...
  /organization/update-event:
    patch:
      operationId: Update event
      parameters:
      - description: event title
        in: query
        name: title
        required: true
        type: string
      - description: 'organizationName: '
        in: query
        name: organizationName
        required: true
        type: string
      - description: update Event with new body
        in: body
        name: UpdateEvent
        required: true
        schema:
          $ref: '#/definitions/domain.Events'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: update event
      tags:
      - Organization
...
  /user/create-event:
    post:
      operationId: Create event from user
      parameters:
      - description: Create event
        in: body
        name: CreateEvent
        required: true
        schema:
          $ref: '#/definitions/domain.Events'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create event
      tags:
      - User
...
  /user/geteventbytitle:
    get:
      operationId: Get event by id
      parameters:
      - description: 'Title: '
        in: query
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      summary: delete event
      tags:
      - User
...
  /user/list/approved-events:
    get:
      operationId: list all approved events
      parameters:
      - description: 'Page number: '
        in: query
        name: page
        required: true
        type: string
      - description: 'Page capacity : '
        in: query
        name: pagesize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      summary: list all approved upcoming events
      tags:
      - User
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"